--- servicetalk-http-netty/src/main/java/io/servicetalk/http/netty/AbstractFilterableStreamingHttpLoadBalancedConnection.java
@@ -1,88 +1,0 @@
-/*
- * Copyright © 2020 Apple Inc. and the ServiceTalk project authors
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package io.servicetalk.http.netty;
-
-import io.servicetalk.concurrent.api.Completable;
-import io.servicetalk.concurrent.api.Publisher;
-import io.servicetalk.concurrent.api.Single;
-import io.servicetalk.http.api.FilterableStreamingHttpConnection;
-import io.servicetalk.http.api.FilterableStreamingHttpLoadBalancedConnection;
-import io.servicetalk.http.api.HttpConnectionContext;
-import io.servicetalk.http.api.HttpEventKey;
-import io.servicetalk.http.api.HttpExecutionContext;
-import io.servicetalk.http.api.HttpExecutionStrategy;
-import io.servicetalk.http.api.HttpRequestMethod;
-import io.servicetalk.http.api.StreamingHttpRequest;
-import io.servicetalk.http.api.StreamingHttpResponse;
-import io.servicetalk.http.api.StreamingHttpResponseFactory;
-
-import static java.util.Objects.requireNonNull;
-
-abstract class AbstractFilterableStreamingHttpLoadBalancedConnection
-        implements FilterableStreamingHttpLoadBalancedConnection {
-
-    private final FilterableStreamingHttpConnection delegate;
-
-    AbstractFilterableStreamingHttpLoadBalancedConnection(final FilterableStreamingHttpConnection delegate) {
-        this.delegate = requireNonNull(delegate);
-    }
-
-    @Override
-    public HttpConnectionContext connectionContext() {
-        return delegate.connectionContext();
-    }
-
-    @Override
-    public <T> Publisher<? extends T> transportEventStream(final HttpEventKey<T> eventKey) {
-        return delegate.transportEventStream(eventKey);
-    }
-
-    @Override
-    public Single<StreamingHttpResponse> request(final HttpExecutionStrategy strategy,
-                                                 final StreamingHttpRequest request) {
-        return delegate.request(strategy, request);
-    }
-
-    @Override
-    public HttpExecutionContext executionContext() {
-        return delegate.executionContext();
-    }
-
-    @Override
-    public StreamingHttpResponseFactory httpResponseFactory() {
-        return delegate.httpResponseFactory();
-    }
-
-    @Override
-    public Completable onClose() {
-        return delegate.onClose();
-    }
-
-    @Override
-    public Completable closeAsync() {
-        return delegate.closeAsync();
-    }
-
-    @Override
-    public Completable closeAsyncGracefully() {
-        return delegate.closeAsyncGracefully();
-    }
-
-    @Override
-    public StreamingHttpRequest newRequest(final HttpRequestMethod method, final String requestTarget) {
-        return delegate.newRequest(method, requestTarget);
-    }
-}
